CircleAvatar account picker modal in AppBar Fixes #2
